+++ glibc/argp/argp_parse.c
@@ -15,7 +15,7 @@ const char argp_prog_doc[] =
 
 static const struct argp_option opts[] = {
 	{ "interface", 'i', "INTERFACE", 0, "Network interface to attach" },
-	{ "verbose", 'v', "VERBOSE", 1, "Display detail" },
+	{ "verbose", 'v', NULL, 1, "Display detail" },
 	{},
 };